+++ lib/distillery/releases/assembler.ex
@@ -4,7 +4,6 @@ defmodule Distillery.Releases.Assembler do
   struct. It creates the release directory, copies applications, and generates release-specific
   files required by `:systools` and `:release_handler`.
   """
-  alias Config.Reader
   alias Distillery.Releases.Conf
   alias Distillery.Releases.Release
   alias Distillery.Releases.Environment
@@ -611,7 +610,7 @@ defmodule Distillery.Releases.Assembler do
     end
   end
 
-  # Generated when Mix.Config provider is active, default + provided sys.config
+  # Generated when Mix.Configig provider is active, default + provided sys.config
   defp generate_sys_config(%Release{profile: %Profile{} = profile} = rel) do
     overlay_vars = profile.overlay_vars
     config_exs_path = profile.config
@@ -656,7 +655,7 @@ defmodule Distillery.Releases.Assembler do
                {:ok, tokens, _} <- :erl_scan.string(String.to_charlist(templated)),
                {:ok, sys_config} <- :erl_parse.parse_term(tokens),
                :ok <- validate_sys_config(sys_config),
-               merged <- Release.merge(base_config, sys_config) do
+               merged <- Config.Reader.merge(base_config, sys_config) do
             merged
           else
             err ->